+++ Ascension/src/blockfighter/shared/Globals.java
@@ -767,6 +767,7 @@ public class Globals {
     public static int SERVER_WIN_KILL_COUNT = 30;
     public static int SERVER_MATCH_DURATION = 300000;
     public static int SERVER_ROOM_LEVEL_DIFF = 2;
+    public static int SERVER_MATCH_TIME_REMAINING_THRESHOLD = 90000;
 
     public static int HUB_SERVER_TCP_PORT = 25566;
     public static String CLIENT_HUB_SERVER_ADDRESS = "asc-hub.servegame.com";
@@ -1218,6 +1219,9 @@ public class Globals {
             if (prop.getProperty("leveldiff") != null) {
                 SERVER_ROOM_LEVEL_DIFF = Integer.parseInt(prop.getProperty("leveldiff"));
             }
+            if (prop.getProperty("matchtimeremainingthreshold") != null) {
+                SERVER_MATCH_TIME_REMAINING_THRESHOLD = Integer.parseInt(prop.getProperty("matchtimeremainingthreshold"));
+            }
         } catch (final FileNotFoundException e) {
             log(Globals.class,
                     "Config", "config.properties not found in root directory. Using default server values.", Globals.LOG_TYPE_DATA);
@@ -1262,6 +1266,8 @@ public class Globals {
                     "Config", "Match Duration(ms): " + SERVER_MATCH_DURATION, Globals.LOG_TYPE_DATA);
             log(Globals.class,
                     "Config", "Room Level Difference(±): " + SERVER_ROOM_LEVEL_DIFF, Globals.LOG_TYPE_DATA);
+            log(Globals.class,
+                    "Config", "Minimum Remaining Time for joinable Room: " + SERVER_MATCH_TIME_REMAINING_THRESHOLD, Globals.LOG_TYPE_DATA);
 
         }
     }